+++ PrebidMobile/AdUnits/Native/NativeRequest.swift
@@ -110,77 +110,3 @@ public class NativeRequest: AdUnit {
         }
     }
 }
-
-public class ContextType: SingleContainerInt {
-    
-    @objc
-    public static let Content = ContextType(1)
-    
-    @objc
-    public static let Social = ContextType(2)
-    
-    @objc
-    public static let Product = ContextType(3)
-    
-    @objc
-    public static let Custom = ContextType(500)
-    
-}
-
-public class ContextSubType: SingleContainerInt {
-    @objc
-    public static let General = ContextSubType(10)
-
-    @objc
-    public static let Article = ContextSubType(11)
-
-    @objc
-    public static let Video = ContextSubType(12)
-
-    @objc
-    public static let Audio = ContextSubType(13)
-
-    @objc
-    public static let Image = ContextSubType(14)
-
-    @objc
-    public static let UserGenerated = ContextSubType(15)
-
-    @objc
-    public static let Social = ContextSubType(20)
-
-    @objc
-    public static let email = ContextSubType(21)
-
-    @objc
-    public static let chatIM = ContextSubType(22)
-
-    @objc
-    public static let SellingProduct = ContextSubType(30)
-
-    @objc
-    public static let AppStore = ContextSubType(31)
-
-    @objc
-    public static let ReviewSite = ContextSubType(32)
-
-    @objc
-    public static let Custom = ContextSubType(500)
-}
-
-public class PlacementType: SingleContainerInt {
-    @objc
-    public static let FeedContent = PlacementType(1)
-
-    @objc
-    public static let AtomicContent = PlacementType(2)
-
-    @objc
-    public static let OutsideContent = PlacementType(3)
-
-    @objc
-    public static let RecommendationWidget = PlacementType(4)
-
-    @objc
-    public static let Custom = PlacementType(500)
-}
